+++ addons/curved_lines_2d/scalable_vector_shape_2d.gd
@@ -57,7 +57,7 @@ enum ShapeType {
 	## Provides one handle to change [member size]. The [member size] determines the radii of the
 	## ellipse on the y- and x- axis, so [member rx] and [member ry] are always sync'ed with
 	## [member size] (and vice-versa)
-		## The [member offset] can change by using the pivot-tool in the 2D Editor
+	## The [member offset] can change by using the pivot-tool in the 2D Editor
 	ELLIPSE
 }
 
@@ -69,6 +69,18 @@ enum CollisionObjectType {
 	RIGID_BODY_2D,
 	CHARACTER_BODY_2D,
 	PHYSICAL_BONE_2D
+}
+
+enum StrokeExtrusionDirection {
+	## The default for a stroke is to draw outward of its defined polyline in both directions
+	## So if the [member stroke_width] is 8px, it draws 4px inward and 4px outward
+	MIDDLE,
+	## If the stroke extrusion direction is set to outward, it will be drawn outside its polyline points
+	## So if the [member stroke_width] is 8px, it draws 0px inward and 8px outward
+	OUTWARD,
+	## If the stroke extrusion direction is set to inward, it will be drawn inside its polyline points
+	## So if the [member stroke_width] is 8px, it draws 8px inward and 0px outward
+	INWARD
 }
 
 @export_group("Fill")
@@ -144,6 +156,12 @@ var stroke_width := 10.0:
 		line_joint_mode = _ljm
 		if is_instance_valid(line):
 			line.joint_mode = _ljm
+		assigned_node_changed.emit()
+
+## The extrusion direction of the stroke. Only applies for closed shapes.
+@export var extrusion_direction := StrokeExtrusionDirection.MIDDLE:
+	set(_ed):
+		extrusion_direction = _ed
 		assigned_node_changed.emit()
 
 ## The 'Stroke' of a [ScalableVectorShape2D] is simply an instance of a [Line2D] node
@@ -519,12 +537,31 @@ func _notification(what: int) -> void:
 	if what == NOTIFICATION_LOCAL_TRANSFORM_CHANGED or what == NOTIFICATION_TRANSFORM_CHANGED:
 		transform_changed.emit(self)
 	if what == NOTIFICATION_EDITOR_PRE_SAVE:
-		if is_instance_valid(skeleton):
-			for i in skeleton.get_bone_count():
-				skeleton.get_bone(i).apply_rest()
-			if is_instance_valid(bone):
-				global_position = bone.global_position
-				global_rotation = bone.global_rotation
+		reset_skeleton_to_rest_pose()
+		_prune_unused_colliders_and_lines()
+
+
+func _prune_unused_colliders_and_lines():
+	if is_instance_valid(line):
+		for ch in line.get_children():
+			if ch is Line2D and not ch.visible:
+				line.remove_child(ch)
+				ch.free()
+	if is_instance_valid(collision_object):
+		for ch in collision_object.get_children():
+			if ch is CollisionPolygon2D and ch.disabled and not ch.visible:
+				collision_object.remove_child(ch)
+				ch.free()
+
+
+func reset_skeleton_to_rest_pose():
+	if is_instance_valid(skeleton):
+		for i in skeleton.get_bone_count():
+			skeleton.get_bone(i).apply_rest()
+		if is_instance_valid(bone):
+			global_position = bone.global_position
+			global_rotation = bone.global_rotation
+
 
 func _on_dimensions_changed():
 	if shape_type == ShapeType.RECT:
@@ -744,8 +781,11 @@ func _update_assigned_nodes(polygon_points : PackedVector2Array) -> void:
 
 	if (is_instance_valid(poly_stroke) or (is_instance_valid(line) and is_instance_valid(collision_object)) or (is_instance_valid(line) and is_instance_valid(navigation_region))) and not cached_outline.size() < 2:
 		var cap_mode := Geometry2D.END_JOINED if is_curve_closed() else CAP_MODE_MAP[begin_cap_mode]
-		var result := Geometry2DUtil.calculate_polystroke(cached_outline,
-				stroke_width * 0.5, cap_mode, JOINT_MODE_MAP[line_joint_mode])
+		var result := Geometry2DUtil.calculate_polystroke(
+				cached_outline, stroke_width * 0.5,
+				cap_mode, JOINT_MODE_MAP[line_joint_mode],
+				_get_stroke_extrusion(cached_outline)
+		)
 		cached_poly_strokes = result
 		# add to list of updated collision polygons
 		if is_instance_valid(collision_object):
@@ -760,7 +800,10 @@ func _update_assigned_nodes(polygon_points : PackedVector2Array) -> void:
 		navigation_polygons.append(polygon_points)
 
 	if is_instance_valid(line):
-		line.points = polygon_points
+		if not extrusion_direction == StrokeExtrusionDirection.MIDDLE and is_curve_closed():
+			line.points = _get_stroke_points_with_extrusion(polygon_points)
+		else:
+			line.points = polygon_points
 		line.closed = is_curve_closed()
 	if is_instance_valid(poly_stroke):
 		var polygon_indices : Array = []
@@ -812,6 +855,23 @@ func _update_polygon_texture(poly := polygon, grow := false):
 				poly.texture_scale = poly.texture.get_size() / box.size
 
 
+func _get_stroke_extrusion(points : PackedVector2Array, is_hole := false) -> float:
+	if extrusion_direction == StrokeExtrusionDirection.MIDDLE or not is_curve_closed():
+		return 0.0
+	var offs := -stroke_width * 0.5 if extrusion_direction == StrokeExtrusionDirection.INWARD else stroke_width * 0.5
+	if is_hole:
+		return -offs
+	return offs
+
+
+func _get_stroke_points_with_extrusion(pts : PackedVector2Array, is_hole := false) -> PackedVector2Array:
+	var extrusion := _get_stroke_extrusion(pts, is_hole)
+	if is_zero_approx(extrusion):
+		return pts
+	var extruded_result := Geometry2D.offset_polygon(pts, extrusion, JOINT_MODE_MAP[line_joint_mode])
+	return pts if extruded_result.is_empty() else extruded_result[0]
+
+
 func _update_assigned_nodes_with_clips(polygon_points : PackedVector2Array, valid_clip_paths : Array[ScalableVectorShape2D]) -> void:
 
 	var merges := valid_clip_paths.filter(func(cp : ScalableVectorShape2D): return cp.use_union_in_stead_of_clipping)
@@ -837,9 +897,12 @@ func _update_assigned_nodes_with_clips(polygon_points : PackedVector2Array, vali
 				[]
 		)
 		var polystroke_result : Array[PackedVector2Array] = []
-		for polyline in cutout_result_polylines:
+		for i in cutout_result_polylines.size():
+			var polyline := cutout_result_polylines[i]
 			polystroke_result.append_array(Geometry2DUtil.calculate_polystroke(polyline,
-					stroke_width * 0.5, Geometry2D.END_JOINED, JOINT_MODE_MAP[line_joint_mode]))
+					stroke_width * 0.5, Geometry2D.END_JOINED, JOINT_MODE_MAP[line_joint_mode],
+					_get_stroke_extrusion(polyline, i > 0)
+			))
 		intersect_results_polystroke = Geometry2DUtil.apply_clips_to_polygon(
 			polystroke_result,
 			Array(clips.map(_clip_path_to_local), TYPE_PACKED_VECTOR2_ARRAY, "", null),
@@ -872,7 +935,7 @@ func _update_assigned_nodes_with_clips(polygon_points : PackedVector2Array, vali
 		else:
 			var polylines := Geometry2DUtil.calculate_outlines(cached_clipped_polygons.duplicate())
 			line.show()
-			line.points = polylines.pop_front()
+			line.points = _get_stroke_points_with_extrusion(polylines.pop_front())
 			# FIXME: closes the loop when original line is not closed
 			line.closed = true
 			var existing = line.get_children().filter(func(c): return c is Line2D)
@@ -882,7 +945,7 @@ func _update_assigned_nodes_with_clips(polygon_points : PackedVector2Array, vali
 			for polyline_index in polylines.size():
 				if polyline_index >= existing.size():
 					existing.append(_make_new_line_2d())
-				existing[polyline_index].points = polylines[polyline_index]
+				existing[polyline_index].points = _get_stroke_points_with_extrusion(polylines[polyline_index], true)
 				existing[polyline_index].width = line.width
 				existing[polyline_index].begin_cap_mode = line.begin_cap_mode
 				existing[polyline_index].end_cap_mode = line.end_cap_mode
@@ -1355,6 +1418,18 @@ func get_subdivided_curve() -> Curve2D:
 			new_curve.set_point_out((i * 2) - 1, segment.get_point_out(1))
 			new_curve.set_point_in(i * 2, segment.get_point_in(2))
 	return new_curve
+
+
+func curve_to_local(curve : Curve2D) -> Curve2D:
+	var c1 := Curve2D.new()
+	for i in curve.point_count:
+		var pos := to_local(curve.get_point_position(i))
+		var abs_in := to_local(curve.get_point_position(i) + curve.get_point_in(i))
+		var abs_out := to_local(curve.get_point_position(i) + curve.get_point_out(i))
+		c1.add_point(pos)
+		c1.set_point_in(i, abs_in - pos)
+		c1.set_point_out(i, abs_out - pos)
+	return c1
 
 
 # Adapted from the GodSVG repository to draw arc in stead of determine bounding box.