+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 2.32.0 - Redemption
+
+## Added
+
+- The Knife tool without "vibes"! <3
+- New property `extrusion_direction` for strokes. If the outline of a shape is closed, the stroke can now also be draw inside and outside of the shape's outline.
+- Button to extract ScalableVectorShape2D from `Line2D`, `Polygon2D` and `CollisionPolygon2D`
+
+## Changed
+
+- Fixed multiple bugs with Undo/Redo by handling state of mouse hover while dragging better
+- Pencil tool now also uses `BasicFit`'s curve fitting algorithm to reduce points and introduce control points (just like the polygon brush)
+- Bugfix: pencil tool can now draw lines of 2 or less straight segments.
+- Bugfix: BasicFit correctly identifies looped and unlooped shapes
+
 ## 2.28.2
 
 ### Added